feat: 생명 사용 및 삭제 로직 + 룸 연동 #BACK-302

Closes BACK-302

See merge request s09-webmobile1-sub2/S09P12A307!66

+++ backend/src/main/java/signiel/heartsigniel/model/member/dto/SignResponse.java
@@ -44,7 +44,7 @@ public class SignResponse {
 
     private List<Authority> roles = new ArrayList<>();
 
-    private int remainLife;
+    private Long remainLife;
 
     private String token;
 }
